# Conflicts:
#	CMakeLists.txt
#	Common/ac/gamesetupstructbase.cpp
#	Common/ac/gamestructdefines.h
#	Common/core/assetmanager.cpp
#	Common/core/def_version.h
#	Common/util/file.cpp
#	Editor/AGS.Editor/Resources/agsdefns.sh
#	Editor/AGS.Editor/app.manifest
#	Editor/AGS.Types/Enums/ScriptAPIVersion.cs
#	Editor/AGS.Types/Properties/AssemblyInfo.cs
#	Engine/ac/draw.cpp
#	Engine/ac/drawingsurface.cpp
#	Engine/ac/file.cpp
#	Engine/ac/global_character.cpp
#	Engine/ac/global_character.h
#	Engine/ac/global_game.cpp
#	Engine/ac/global_object.cpp
#	Engine/ac/object.cpp
#	Engine/ac/overlay.cpp
#	Engine/ac/overlay.h
#	Engine/ac/path_helper.h
#	Engine/ac/room.cpp
#	Engine/ac/roomobject.cpp
#	Engine/ac/roomobject.h
#	Engine/ac/system.cpp
#	Engine/game/game_init.cpp
#	Engine/main/game_run.cpp
#	version.json

+++ libsrc/mojoAL/mojoal.c
@@ -531,7 +531,7 @@ struct ALCcontext_struct
 };
 
 /* forward declarations */
-static int source_get_offset(ALsource *src, ALenum param);
+static float source_get_offset(ALsource *src, ALenum param);
 static void source_set_offset(ALsource *src, ALenum param, ALfloat value);
 
 /* the just_queued list is backwards. Add it to the queue in the correct order. */
@@ -3940,9 +3940,9 @@ static void source_pause(ALCcontext *ctx, const ALuint name)
     }
 }
 
-static int source_get_offset(ALsource *src, ALenum param)
+static float source_get_offset(ALsource *src, ALenum param)
 {
-    int offset = 0;
+    float offset = 0.0f;
     int framesize = sizeof(float);
     int freq = 1;
     if (src->type == AL_STREAMING) {
@@ -3965,7 +3965,7 @@ static int source_get_offset(ALsource *src, ALenum param)
         case AL_SAMPLE_OFFSET: return offset / framesize; break;
         case AL_SEC_OFFSET: return (offset / framesize) / freq; break;
         case AL_BYTE_OFFSET: return offset; break;
-        default: return 0; break;
+        default: return 0.0f; break;
     }
 }
